+++ blog/services/UserService.java
@@ -11,9 +11,9 @@ import blog.payloads.UserDto;
 public interface UserService {
 	
 	UserDto createUser(UserDto user);
-	UserDto updateUser(UserDto user, Long userId);
-	UserDto getUserById(Long userId);
+	UserDto updateUser(UserDto user, Integer userId);
+	UserDto getUserById(Integer userId);
     List<UserDto> getAllUsers();
-    void deleteUser(Long userId);
+    void deleteUser(Integer userId);
 
 }
